chore(tests): move writer tests under tests/Writers/ for symmetry with Readers/
